Adding Virtual Atom for pescan/emass.
Need first to add python bindings for centers, such that microstrain and
friends is available.
Then implement in python a VA class which takes a vff and an escan
functional. This class should be able to print escan input for
fractional occupations. When computing derivatives, it should reuse the
wavefunctions from a previous call. Finally numpy.linalg.lsq can be used to
interpolate for a virtual atom derivatives.

Can we do all this directly in fortran!!