Crays fcntl making life difficult.
  * changed flock for lockf
  * moved lock from file itself to separate lockfile
  * added acquire_lock context
As usual cray's make life difficult. flock does not work on
compute nodes when using aprun. lockf does but requires the file to be
opened with "w" only! Hence we lock a second file rather than the on of
interest. We can't delete either, in case another process is competing
with this one. Inconvenient to say the least.